requirements.txt
Flask==2.3.2
flask-smorest
Flask-SQLAlchemy
Flask-Migrate
passlib
marshmallow
python-dotenv
gunicorn
psycopg2==2.9.1